third_party/blink/web_tests/svg/animations/use-animate-width-and-height-expected.txt

SVG 1.1 dynamic animation tests

Test animation of use element width/height

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS use.width.animVal.value is 100
PASS use.width.baseVal.value is 100
PASS use.height.animVal.value is 100
PASS use.height.baseVal.value is 100
PASS use.getAttribute('width') is '100'
PASS use.getAttribute('height') is '100'
PASS shadowRoot.firstChild.width.animVal.value is 100
PASS shadowRoot.firstChild.height.animVal.value is 100
PASS use.width.animVal.value is 105
PASS use.width.baseVal.value is 100
PASS use.height.animVal.value is 105
PASS use.height.baseVal.value is 100
PASS use.getAttribute('width') is '100'
PASS use.getAttribute('height') is '100'
PASS shadowRoot.firstChild.width.animVal.value is 105
PASS shadowRoot.firstChild.height.animVal.value is 105
PASS use.width.animVal.value is 115
PASS use.width.baseVal.value is 100
PASS use.height.animVal.value is 115
PASS use.height.baseVal.value is 100
PASS use.getAttribute('width') is '100'
PASS use.getAttribute('height') is '100'
PASS shadowRoot.firstChild.width.animVal.value is 115
PASS shadowRoot.firstChild.height.animVal.value is 115
PASS use.width.animVal.value is 125
PASS use.width.baseVal.value is 100
PASS use.height.animVal.value is 125
PASS use.height.baseVal.value is 100
PASS use.getAttribute('width') is '100'
PASS use.getAttribute('height') is '100'
PASS shadowRoot.firstChild.width.animVal.value is 125
PASS shadowRoot.firstChild.height.animVal.value is 125
PASS use.width.animVal.value is 135
PASS use.width.baseVal.value is 100
PASS use.height.animVal.value is 135
PASS use.height.baseVal.value is 100
PASS use.getAttribute('width') is '100'
PASS use.getAttribute('height') is '100'
PASS shadowRoot.firstChild.width.animVal.value is 135
PASS shadowRoot.firstChild.height.animVal.value is 135
PASS successfullyParsed is true

TEST COMPLETE
